feature: URL Query handling
Previously GPget would naively just append the extention to the URL
failing to note if there were query parameters or fragments at the
end.  This fixes that behavior though it should be noted that for
the extension the *SAME* query parameters will be passed along.

If you need different query parameters this begins to be tread
beyond the scope of this project and one is likely better off
scripting this behavior.