feature: Clearsign support
This one is a bit wonky, so consider this a prototype at the monent
as I want to re-work most of the error handling.  At the moment you
need to know if the remote endpoint is clearsigned or not.  If it's
clearsigned then everything is hunky dory by adding the `-c` flag.
On the other hand if you try to pull down a detached signature file
(`.asc`) with the clearsign flag it will error out with a _very_
nuanced error about being a SIGNATURE not a SIGNED MESSAGE.